requirements.txt
langchain
langchain-google-genai
pydantic
python-docx
google-cloud
google-api-core
google-cloud-aiplatform
dotenv